pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxmaestro"
version = "0.1.0"
description = "YAML-defined voice agent conductor — state machine with mid-turn tool bridging and 3-phase handoff protocol"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
langfuse = ["langfuse>=2.0"]
pipecat = ["pipecat-ai>=0.0.30"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/voxmaestro"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = true
